+++ extensions-core/multi-stage-query/src/main/java/org/apache/druid/msq/indexing/destination/MSQTerminalStageSpecFactory.java
@@ -17,21 +17,15 @@
  * under the License.
  */
 
-package org.apache.druid.msq.guice;
+package org.apache.druid.msq.indexing.destination;
 
-import org.apache.druid.msq.indexing.destination.SegmentGenerationStageSpec;
-import org.apache.druid.msq.indexing.destination.TerminalStageSpec;
 import org.apache.druid.sql.calcite.planner.PlannerContext;
 import org.apache.druid.sql.calcite.rel.DruidQuery;
 
-/**
- * Configures ingestion queries to create new segments with the results in all cases.
- */
-public class SegmentGenerationTerminalStageSpecFactory implements MSQTerminalStageSpecFactory
+public interface MSQTerminalStageSpecFactory
 {
-  @Override
-  public TerminalStageSpec createTerminalStageSpec(DruidQuery druidQuery, PlannerContext plannerContext)
-  {
-    return SegmentGenerationStageSpec.instance();
-  }
+  /**
+   * Creates a {@link TerminalStageSpec} which determines the final of a query.
+   */
+  TerminalStageSpec createTerminalStageSpec(DruidQuery druidQuery, PlannerContext plannerContext);
 }
